rework: Organize files in /handlers
