+++ server/src/router/routes.js
@@ -3,14 +3,14 @@ import router from "./router.js";
 import auth from "../middleware/auth.js";
 import allow from "../middleware/allow.js";
 
-import { register, login, whoami, update } from "../handlers/auth.js";
+import { register, login, whoami, update } from "../handlers/users.js";
 import {
     getQuizzes,
     createQuiz,
     getQuiz,
     updateQuiz,
     deleteQuiz,
-} from "../handlers/quiz.js";
+} from "../handlers/quizzes.js";
 
 // Public routes
 router.register(async (r) => {
